ability to remove filter chips
