+++ src/app/components/manga-filter-form/manga-filter-form.component.html
@@ -45,26 +45,36 @@
 
 </form>
 
-<!-- <pre>
-  {{QueryFor
-    mGroup.value | json }}
-</pre> -->
+<pre>
+  {{QueryFormGroup.value | json }}
+</pre>
+
 <mat-chip-list aria-label="Query Chips">
-  <mat-chip *ngIf="this.mangaSearchTerm?.value!==''" color="primary" selected> Search: {{this.mangaSearchTerm?.value}}
+
+  <mat-chip (click)="this.mangaSearchTerm?.setValue('')" *ngIf="this.mangaSearchTerm?.value!==''" color="primary"
+    selected> Search: {{this.mangaSearchTerm?.value}}
+    <mat-icon>close</mat-icon>
   </mat-chip>
-  <mat-chip *ngIf="this.mangaGenre?.value?.length" color="primary" selected>
+
+  <mat-chip (click)="this.mangaGenre?.setValue([])" *ngIf="this.mangaGenre?.value?.length" color="primary" selected>
     <form-genre-display [mangaGenre]="this.mangaGenre">
     </form-genre-display>
-  </mat-chip>
-  <mat-chip *ngIf="this.mangaStatus?.value" color="primary" selected>Status: {{this.mangaStatus?.value | titlecase}}
+    <mat-icon>close</mat-icon>
   </mat-chip>
 
-  <mat-chip *ngIf="this.mangaOrderBy?.value" color="primary" selected> Order by: {{this.mangaOrderBy?.value | titlecase
-    |
-    snakeCaseToSpace}}</mat-chip>
+  <mat-chip (click)="this.mangaStatus?.setValue(null)" *ngIf="this.mangaStatus?.value" color="primary" selected>Status:
+    {{this.mangaStatus?.value | titlecase}}
+    <mat-icon>close</mat-icon>
+  </mat-chip>
+
+  <mat-chip (click)="this.mangaOrderBy?.setValue(null)" *ngIf="this.mangaOrderBy?.value" color="primary" selected> Order
+    by: {{this.mangaOrderBy?.value | titlecase | snakeCaseToSpace}}
+    <mat-icon>close</mat-icon>
+  </mat-chip>
 
   <mat-chip *ngIf="formHasChanged" (click)="clearQueryValues()" color="accent">
+    Clear all Filters
     <mat-icon>close</mat-icon>
-    Reset filters
   </mat-chip>
+
 </mat-chip-list>
